requirements.txt
graphviz>=0.8.4
Keras>=2.2.0
lxml>=4.1.1
Markdown>=2.6.9
matplotlib>=2.2.3
numpy>=1.14.5
opencv-python>=3.4.0.12
pandas>=0.21.1
Pillow>=4.3.0
tensorboard>=1.6.0
tensorflow>=1.3.0
tensorflow-gpu>=1.4.0
tqdm>=4.11.2
albumentations>=0.1.2